+++ lib/src/templates/default/web/main.dart
@@ -2,5 +2,5 @@ import 'package:pulsar_web/pulsar.dart';
 import 'views/counter/counter.dart';
 
 void main() {
-  runApp(CounterApp());
+  mountApp(CounterApp(), selector: "#app");
 }