+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.0
+- Support for the newest version of Pulsar 0.4.0.
+- Breaking changes to the templates for create command.
+
 ## 0.0.15
 - Added support for the newest version of Pulsar 0.3.1.